v.0.5.0 The Twitter API 1.1 edition
Hi, grab a drink, maybe a snack, this one is kind of long.

As of March 5th, 2013, Twitter is officially removing support for their
v1 API. As part of this, they are also removing support for
unauthenticated requests to their API (yadda yadda yadda politics,
advertising, big money).

Until now, grailbird_updater used API v1 to make unauthenticated
requests to update tweets, since it can no longer do this,
grailbird_updater will now prompt users ONCE per archive in order to
authenticate it to retrieve tweets from that account.

PLEASE NOTE: The warnings are to be taken seriously, the YAML key files
should not be stored on public servers. See LICENSE.txt for more about
the lack of warranty or liability provided.

There is also support for CSVs as of this version, it is optional. To
make the best use of it, redownload your archive from Twitter and start
updating that archive with the CSV flag option turned on.

Key commits in version 0.5.0 are:
* 4594e27257 API v1.1
* e604f73548 ad27ad3e97 c48968bee2 Rudimentary CSV support